# rivulet-server/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      role:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.GetSessionResponse:
    properties:
      account_id:
        type: string
      role:
        type: string
      sessions:
        items:
          $ref: '#/definitions/models.GetSessionResponseItem'
        type: array
    type: object
  models.GetSessionResponseItem:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      device_name:
        type: string
      expires_at:
        type: string
      family_id:
        type: string
      ip_address:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API for Rivulet
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: Rivulet API
  version: "0.1"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login into an Admin or User account. This endpoint is unauthenticated.
      parameters:
      - description: Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login into an Admin or User account.
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout from an Admin or User account. This endpoint is authenticated.
      parameters:
      - description: Logout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout from an Admin or User account.
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh an Admin or User account's Access and Refresh Tokens. This
        endpoint is unauthenticated.
      parameters:
      - description: Refresh
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Refresh an Admin or User account's Access and Refresh Tokens.
      tags:
      - Auth
  /api/v1/auth/session:
    delete:
      consumes:
      - application/json
      description: End a session for an Admin or User account. This endpoint is authenticated.
        If the caller is Admin, they can end sessions for both Admin and User Accounts.
        If the caller is User, they can only end their own sessions.
      parameters:
      - description: Family ID
        in: query
        name: family_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: End a session for an Admin or User account.
      tags:
      - Auth
  /api/v1/auth/session/all:
    delete:
      consumes:
      - application/json
      description: End all sessions for an Admin or User account. This endpoint is
        authenticated. If the caller is Admin, they can end sessions for both Admin
        and User Accounts. If the caller is User, they can only end their own sessions.
      parameters:
      - description: Account ID
        in: query
        name: account_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: End all sessions for an Admin or User account.
      tags:
      - Auth
    get:
      consumes:
      - application/json
      description: Get active sessions for an Admin or User account. This endpoint
        is authenticated. If the caller is Admin, the `account_id` query parameter
        is required. If the caller is User, the `account_id` query parameter is ignored.
      parameters:
      - description: Account ID
        in: query
        name: account_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active sessions for an Admin or User account.
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"